doit: Add trace command
The Aspeed AHB Bus Controller features a command recording facility
for tracing MMIO accesses to a designated register.

The new `trace` subcommand of doit exposes the command recording
facility. Tracing terminates upon reception of SIGINT, at which point the
command dumps the trace buffer to stdout.

Currently support is limited to the AST2600 due to handling of the SoC
memory map. The implementation uses the SRAM as the trace buffer as it
is otherwise unused by Linux, but the location of the SRAM in the
address space varies between SoC generations. Hopefully at some point we
improve our address-space layout handling to lift this limitation.

Here's an example invocation where `echo hello\n` is remotely written to the
BMC console, resulting in AHB reads on the BMC-side:

```
root@ast2600:~# ./doit trace 0x1e784000 1:0 read > trace
[*] Probing local interfaces
[*] Probing AHB interfaces
[*] Performing interface discovery via devmem
[*] Detected AST2600 A1
[*] No description for the PCIe configuration layout on the AST2600 A1
[*] Detected devmem interface
[*] Initialising devmem interface
[*] Initialised devmem AHB interface
[*] Started AHB trace for 0x1e784000
^C[*] Stopped AHB trace
.
root@ast2600:~# hexdump -C trace
00000000  65 63 68 6f 20 68 65 6c                           |echo hel|
00000008
```

Note that the data is truncated: The command recording facility will not flush
partial word data to the trace buffer, so the tail may be truncated for
captures of byte and half-word accesses.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>