trace: set the offset / style automatically
The trace address probe (i.e. listen) register has the two low bits
reserved so you can't set a listen address that's unaligned. The "style"
register allows configuring the trace engine to only record a specific
1/2 byte segment of that 4 byte listening area.

There's no real need to force the user to specify the correct value for
the "style" / "offset" since it can be derived from the listen address
and access size specified on the command line.

Signed-off-by: Oliver O'Halloran <oliver.ohalloran@oracle.com>