sfc: Always use 32bit MMIOs for reading command results
The P2A bridge has a quirk where it performs a 4-byte read on the AHB
regardless of the read size requested on the PCI side. This is a problem
when accessing the serial flash controller(s) since when in User command
mode each byte of IO to the SFC's MMIO range is converted into a byte of
traffic on the SPI bus. The net result is that we lose part of the
command response data unless we perform 4 byte reads.

For example, the flash chip ID command (0xef) results in a three
byte response. If we read the response with a single 4 byte access
the SFC sees:

Wr 1 - 9f # send the flash ID command
Rd 4 - ef 40 19 00

However, if we attempt to read the response as a series of single
byte reads the SFC sees:

Wr 1 - 9f # send the flash ID command
Rd 4 - ef 40 19 00
Rd 4 - 00 00 00 00
Rd 4 - 00 00 00 00

The P2A extracts the relevant byte from the 4 byte response so assuming
we were doing repeated reads to address 0 we'd see: "ef 00 00" which is
obviously wrong.

Fix this by always using explicit 4 byte reads and manually extracting
the required number of bytes. As far as I can tell this is fine for
read commands since any extra SPI bus traffic won't produce side
effects. It would be a problem for writes, but for writes the P2A appears
to be well behaved.

Signed-off-by: Oliver O'Halloran <oliver.ohalloran@oracle.com>