trace: re-init HW accessor before dumping
For long lived traces the HW bridge being used to access the trace
buffer may have changed state while tracing was occuring. This can
result in reading garbage, e.g.

[*] Found AHBC at 0x1e600000 and SRAM at 0x1e720000
[*] trace_start: 0x00000000 4:0 0
[*] Started AHB trace for 0x00000000
[*] trace_stop: csr: 0xffffffff
[*] Stopped AHB trace
[*] trace_dump: csr: 0xffffffff
[*] trace_dump: buf: 0xffffffff
[*] Ring buffer has wrapped, dumping trace buffer from write pointer at 0xfffffffe for 4

Which is clearly nonsense. In this case the 0xFFs response are because
another invokation of the tool disabled the P2A bridge that was being
used to access the trace registers. Fix this by re-initalising the AHB
bridge that we're using before touching the HW.

Signed-off-by: Oliver O'Halloran <oliver.ohalloran@oracle.com>