README: Add cross compilation example
For those who aren't familiar with meson.

The instructions will produce a binary that is linked against the libc
the toolchain was built with:

 $ readelf -d build-arm/src/culvert  |grep -i needed
 0x00000001 (NEEDED)                     Shared library: [libc.so.6]

Ideally we could instruct meson to statically link, but for now the
following workaround can be used in addition to the cross compliation
instructions:

 meson compile -v -C build-arm

Observe the culvert link line. Note that the libfdt link step for dtc
may come last, so you will have to scroll up.

Cut/paste the gcc invocation and add -static. The resulting binary will
include it's own libc copy, so you can run it on pretty much any
machine:

 $ readelf -d build-arm/src/culvert
 There is no dynamic section in this file.

Signed-off-by: Joel Stanley <joel@jms.id.au>