rev: Fix g6 identification logic
Despite the datasheet saying that the top bits of SCU004 on the
AST2500 should always be either 0b111111 or 0b111101 (depending on
which line of the datasheet you look at), we've observed an AST2500
system with its SCU004 reading 0x000ea088 (possibly only after
executing some errant u-boot code).  In combination with an SCU07C
value of 0x04030303, this led to the previous logic erroneously
identifying it as an AST2600.

To accommodate such cases we now constrain the g6-detection logic to
trigger only if the upper bits of *both* probed registers are clear
(which is expected on g6 hardware), instead of if either of them is.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>