cmd: write: Always enable --live mode
Booting an AST2500 from an erased (or partially erased) SPI flash can
leave it in a state in which culvert is unable to recover it (its
attempts to identify the flash chip fail) if not in --live mode,
possibly due to interference in the SPI controller/bus from uncached
instruction fetches.  Given this possibility and the scenario being a
relatively common one, let's just always enable --live mode (retaining
the flag as a no-op for backwards compatibility).

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>