Move bridge names into struct bridge_driver
Unify the previous ahb_interface_names array into the existing
bridge-specific information in struct bridge_driver.  This also
eliminates enum ahb_bridge, replacing usage of it with pointers to
bridge drivers instead.  For the special-casing of ahb_devmem we now
have a bool flag indicating if the bridge is for local (on-BMC use),
which happens to align with what the code doing those special-case
checks was really interested in anyway.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>