wdt: Use counter restart register to restart counter
The ast2500-wdt hardware requires a write of a magic value (0x4755) to
the counter restart register (0x08) in order to load the current
reload register (0x04) into the counter register (0x00).  Previously
we'd been omitting this step, which caused watchdog-based resets to
use whatever value had been sitting in the counter register,
potentially delaying the reset by a problematic length of time (for
example, enough to allow the existing in-memory firmware to issue
flash writes and corrupt the new firmware written by a 'write
firmware' operation before we reboot into it).

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>