wdt: Wait for timer expiration period before un-gating ARM clock
Allowing the ARM core to resume execution before the reset has
actually fired can lead to undesirable situations such as old
in-memory firmware corrupting a new firmware image just written by a
'write firmware' command.  Add a usleep() call to wdt_perform_reset()
so that it waits the necessary amount of time itself rather than
returning a duration for the caller to wait.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>