ahb: Constrain siphon operations to specified memory region
Particularly, ahb_siphon_in() just kept reading until the FD was
exhausted. This isn't always sensible.

Rework the ahb_siphon_in() prototype to add a size parameter, and limit
the data read from the FD to the size of the memory region. However,
also provide the escape-hatch of passing -1 as the length for unbounded
reads and writes.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>