devcontainer: add manifest and Dockerfile
Devcontainer allows to develop software within a container which
is isolated from the system, except for passing the workspace to
the container.
This allows the development on non-Linux systems such as macOS
without having to run a full-blown virtual machine, but only a small
container (where the Linux layer is either remote or in a small QEMU
instance via e.g. Colima).

Signed-off-by: Tan Siewert <tan@siewert.io>