meson: add aarch64-linux cross-compile config
There are a few systems on the market that have an ASPEED chip, but have
an aarch64 CPU (e.g. Ampere's Mt. Jefferson Reference platform) plus the
AST2700 will have an aarch64 CPU.

Tested: Cross-compile culvert on an x86 machine to aarch64 and run the
binary on an ALTRAD8UD-1L2T system (Ampere Altra Q80-30).

Signed-off-by: Tan Siewert <tan@siewert.io>